Extract id effect into a custom hook
