+++ frontend/src/features/noteTaking/hooks/useNotebookIdEffect.hooks.ts
@@ -1,0 +1,33 @@
+import { useEffect } from "react";
+
+interface UseNotebookIdEffectProps {
+    notebookId: string | undefined;
+    updateCurrentNotebookId: (id: number) => void;
+    clearCurrentNotebookId: () => void;
+}
+
+function useNotebookIdEffect({
+    notebookId,
+    updateCurrentNotebookId,
+    clearCurrentNotebookId,
+}: UseNotebookIdEffectProps) {
+    useEffect(() => {
+        if (notebookId === undefined) {
+            return;
+        }
+
+        const id = Number(notebookId);
+        if (Number.isNaN(id)) {
+            clearCurrentNotebookId();
+            return;
+        }
+
+        updateCurrentNotebookId(id);
+
+        return () => {
+            clearCurrentNotebookId();
+        };
+    }, [notebookId, updateCurrentNotebookId, clearCurrentNotebookId]);
+}
+
+export default useNotebookIdEffect;
